refactor: OAuth2.0 토큰 만료시간 통일, OAuth2.0 로그아웃 기능 구현 (#23)
* refactor: Redis 저장 항목에서 Access Token 제거

* chore: Access Token 만료 시간 추가

* refactor: 토큰 만료 시간 통일

* test: 재로그인 시 만료 시간 재설정 테스트 추가

* chore: Google OAuth2.0 토큰 제거 정보 추가

* feat: Google OAuth2.0 토큰 제거 기능 구현

* chore: kakao OAuth2.0 로그아웃 정보 추가

* refactor: revoke -> logout 명칭 변경

* feat: kakao OAuth2.0 로그아웃 기능 구현

* refactor: 역할에 맞게 메서드명 변경

* test: 테스트에 Token Type 추가

* refactor: 매직 리터럴 상수화

* test: Service Token 픽스처 생성

* test: FakeOAuth 객체 구현

+++ src/main/java/com/dobugs/yologaauthenticationapi/support/dto/response/OAuthTokenDto.java
@@ -1,4 +1,4 @@
 package com.dobugs.yologaauthenticationapi.support.dto.response;
 
-public record OAuthTokenDto(String accessToken, int expiresIn, String refreshToken, int refreshTokenExpiresIn, String tokenType) {
+public record OAuthTokenDto(String accessToken, long expiresIn, String refreshToken, long refreshTokenExpiresIn, String tokenType) {
 }
